pyproject.toml
[project]
name = "disease"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.0",
    "joblib>=1.5.3",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.12",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.49",
    "unicorn>=2.1.4",
    "uvicorn>=0.45.0",
    "xgboost>=3.2.0",
]
[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "database*", "ml*", "scripts*"]
